finally updated

+++ src/main/java/net/aspw/client/util/misc/NewFallingPlayer.kt
@@ -38,7 +38,7 @@ class NewFallingPlayer(
             if (v < 1.0f) {
                 v = 1.0f
             }
-            var fixedJumpFactor = (jumpMovementFactor).toFloat()
+            var fixedJumpFactor = (jumpMovementFactor)
             if (mc.thePlayer.isSprinting) {
                 fixedJumpFactor = (fixedJumpFactor * 1.3).toFloat()
             }
